The PXD000001 data set

Details

The package contains a subset of the files from the ProteomeXchange PXD000001 data set. This package is meant to be used for teaching and demonstration purposes, when no internet access is available. The functions available in this package mimic the behaviour of rpx for the local files.

To obtain a list of available files, use px1files(), which will return a vector of file names. Additional paramters can be used that will be passed to list.files. See below for an example.

The px1get function takes one or several file names, as returned by px1files() as input and copies the file(s) into the working directory. To copy all the files, use px1get("all").

The data of this package has been retrieved using the "scripts/getPXD000001.R" available in the package directory.

References

The data was originally published in Gatto L, Christoforou A. Using R and Bioconductor for proteomics data analysis. Biochim Biophys Acta. 2014 Jan;1844(1 Pt A):42-51. doi: 10.1016/j.bbapap.2013.04.032. Epub 2013 May 18. Review. PubMed PMID: 23692960.

See also

The rpx package, that was used to download the data from the ProteomeXchange repository.

Examples

## Available functions:
library("PXD000001")
ls("package:PXD000001")

#> [1] "px1files" "px1get"  

fls <- px1files()
fls

#> [1] "erwinia_carotovora.fasta"                                              
#> [2] "F063721.dat"                                                           
#> [3] "F063721.dat-mztab.txt"                                                 
#> [4] "PXD000001_mztab.txt"                                                   
#> [5] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML.gz"
#> [6] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid"   

## Original data
library("rpx")
px <- PXDataset("PXD000001")
pxfiles(px)

#>  [1] "F063721.dat"                                                         
#>  [2] "F063721.dat-mztab.txt"                                               
#>  [3] "PRIDE_Exp_Complete_Ac_22134.xml.gz"                                  
#>  [4] "PRIDE_Exp_mzData_Ac_22134.xml.gz"                                    
#>  [5] "PXD000001_mztab.txt"                                                 
#>  [6] "README.txt"                                                          
#>  [7] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzML" 
#>  [8] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzXML"
#>  [9] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzXML"         
#> [10] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.raw"           
#> [11] "erwinia_carotovora.fasta"                                            
#> [12] "generated"                                                           

not_run({
    ## Get one files
    px1get("erwinia_carotovora.fasta")
    ## Get all the data into working dir
    px1get("all")
})